Initial commit
The initial commit of the app. This commit was delayed for a long time,
due to several factors.

This commit has the basic work-flow of the app that includes:
   1. Get user input through the temporary, make-shift form,
   2. Create an instance of the Todo class,
   3. Initiate an empty todos array,
   4. Push the newly created Todo object into the todos array,
   5. Send the tods array into the localStorage,
   6. Retrieve the the existing todos from localStorage if any,

This basic version was designed to work only on the console, and is
there to get the basic work-flow done. The next step will be to create
a UI and render the todos in lists and date order.