Fix Todo creation and some form logic
Some of the form elements like the priority radio buttons (previously
the priority select fields) were not working properly. Decided to
replace the whole priority logic with radio buttons for better exp. And
also made sure that they are now returning proper priority values from
the form.

Made the necessary changes to createTodo module for the implementation
of this new form input logic. The todos now have their priority and date
values properly.

There was another error with the semantics of the little form-display in
the UI where the todos were being displayed. Implemented a -perhaps-
temporary logic to display the todo.title value as the ID and htmlFor
attributes for the checkbox inputs in the todos display. These ID's
end up remaining in the project or might get replaced with a todo ID
logic in the future.