Overhaul several modules extensively
The following modules have been merged:

- displayLists.js and displayTodos.js with render.js (renamed from
renderModule.js),
- leftMenuNav.js with leftMenu.js,
- todos.js with lists.js,

The following modules were removed:

- createTodo.js,
- displayLists.js,
- displayTodos.js,
- leftMenuNav.js,
- lists.js,
- logic.js,
- todos.js

The following modules were renamed:

- createTodo.js to createData.js,
- todos.js to data.js,
- renderModule.js to render.js,

Refactored a variety of code in the abovementioned modules for better
readability, reusability, modularity, DRY code, single responsibility
and decoupling purposes.

Added a List class to process and manage lists and their interactions
with todos. Will be working on that later on.