feat: add reusable button component
k-kelvin's reusable button component is very handy for all buttons for
this project. they are also customizable to fit the figma design almost
perfectly.

furthermore, this commit integrates the previously parsed svg icons into
this button's configuration so all of our current icons are ready to be
used with this button component.

there is currently an issue in the button component where the types are
throwing some warnings. this will be fixed on a later commit.