feat: migrate the `Button` component to typescript
the original author of the Button component was written in JSX and had
the old `PropTypes` support from a previous react version (before v19).
however, the react v19 update removed the support for `PropTypes`, and
now suggests typescript for type checking. the button component was thus
converted to typescript.

along with that, the older icon components were also causing similar
type errors for a while, which was also fixed.