feat: add mobil components & extract `mobileViews`
the `mobileViews` that is used for both as a type declaration & array
to map over to conditionally render `MobileBottomNav` items as well as
the `MobileLayout` main content view.

this was originally a temporal
type declaration. it is not extracted into an array and the type was
derived from that. furthermore, the mobileViews array is now used not
only as a type declaration, but now is also used to conditionally render
the `MobileBottomNav` items.