refactor: relocate ` MainTitle` & add: prop types
the MainTitle was originally rendered once on the `MobileLayout` but was
later refactored with the introduction of routing. this little refactor
was made to compliment that. with this refactor, the `MainTitle` is now
rendered inside each Route's parent component. however, there's an issue
with the "/" route on mobile layout where, since the `NotesList` is
already rendered inside the "/" as well as `SearchPage` & `ArchiveList`,
it causes duplicates. it will be fixed in a later refactor.

such refactors required some prop-drilling, which also necessitated
additional type declarations. this commit also aims to perform that.