feat: add `addNewNote` logic
after a long and tedious prop drilling and refactoring, a very basic
functionality to add new note and pass it to localStorage was added.
however, the logic is not very robust and needs to be improved, namely
it overwrites a single note to storage instead of creating a new one.
this will be fixed tomorrow.