feat: add CREATE operation to `CreateNoteForm`
the form is now operational. it gets the data in the `FormData` state &
calls `setNotes`. the notes state was originally a single note object.
this was changed aswell so no it is an array of notes, effectively
fixing the previous commit's known issue.

a `useEffect` call sets the current notes to the `localStorage`. this is
the temporary storage solution until a proper backend is implemented.

next up, READ & DELETE operations, after which the UPDATE operation
will be implemented.