feat: implement READ operation from local for MVP
this commit implements the read operation while refactoring a bit to
achieve single source of truth for the notes. previously, as a result of
a mistake, there were two sources of truth, one in the `useData` hook &
one in the `App.tsx`. the notes state turned out to be dupe (i know!).
thus, the migration to `data` & `setData` was performed. the old `notes`
state and its setter will be refactored out in the following commit.