fix: an error in READ opr & refactor: `NoteBody`
the READ operation had an issue with the state being updated & wrongly
initialize. this is now fixed and should work correctly.

furthermore, i have realized that the figma design intends the note body
to be just a `CreateNoteForm` instead of a read-only `NoteBody`. even
though i disagree to that, i have decided to refactor the code to
accommodate the design. the `NoteBody` component is now replaced with
`CreateNoteForm` component. this is a breaking change, so i have added a
commit message to indicate this. the component will be renamed to a more
meaningful name in the future.