feat: archive & refactor: generic action handler
archiving notes is now possible via the archive buttons in both mobile &
desktop views. the `NotesList` now dynamically renders the notes based
on the `activeView` state, i.e: if activeView is "/", all notes are
rendered, if it is "/archive", only archived notes are rendered.

there are some brittle logic with the activeView state. it will be fixed
at a later time, perhaps even Soon™.

a generic `handleNoteActions` function is drilled down as prop that will
call the relevant action handler based on the action type. This will
prevent the need drill unnecessary props down the component tree.