fix: an issue with `TagsList` & `searchQuery`
the tag filtering functionality in `MobileLayout` was not triggering the
required re-render, so, although the state was being set, the filtered
notes were never displayed in mobile view. this was fixed by adding
routing to "/tags/:tag" so that now the filtered notes are displayed
correctly in mobile view.

furthermore, the issue with `searchQuery` not being properly reset when
the user navigates without the nav buttons (they were responsible for
setting the `activeView` state) was also fixed. this will further be
improved in the future, with the addition of `useLocation` hook. this
hook will be the single responsible for setting the `activeView` state,
based on the current route path. this way the activeView will be more
consistent.