refactor: the `CreateNoteForm` now checks for note
the note to be rendered inside the CreateNoteForm is now checked inside
an effect and render the selected note if the id exists; and render the
data[0] otherwise. also the title input is only auto-focused if the path
is "/create".

this is done to accomodate future refactors to the CreateNoteForm where
it will be converted into a `NoteBody`-esque display area, which also
happens to render the form component that will later be extracted from
the current CreateNoteForm.